[refactor] separate product service and interface
